+++ TombEngine/Math/Math.h
@@ -1,16 +1,15 @@
 #pragma once
 #include "Math/Constants.h"
-#include "Math/Containers/EulerAngles.h"
-#include "Math/Containers/GameBoundingBox.h"
-#include "Math/Containers/GameVector.h"
-#include "Math/Containers/Pose.h"
-#include "Math/Containers/Vector2i.h"
-#include "Math/Containers/Vector3i.h"
 #include "Math/Geometry.h"
+#include "Math/Legacy.h"
+#include "Math/Objects/EulerAngles.h"
+#include "Math/Objects/GameBoundingBox.h"
+#include "Math/Objects/GameVector.h"
+#include "Math/Objects/Pose.h"
+#include "Math/Objects/Vector2i.h"
+#include "Math/Objects/Vector3i.h"
 #include "Math/Random.h"
 #include "Math/Solvers.h"
-
-#include "Math/Legacy.h"
 
 namespace TEN::Math
 {
