+++ TombEngine/Math/Objects/EulerAngles.cpp
@@ -1,10 +1,10 @@
 #include "framework.h"
-#include "Math/Containers/EulerAngles.h"
+#include "Math/Objects/EulerAngles.h"
 
 #include "Math/Constants.h"
-#include "Math/Containers/AxisAngle.h"
 #include "Math/Geometry.h"
 #include "Math/Legacy.h"
+#include "Math/Objects/AxisAngle.h"
 
 using namespace TEN::Math;
