+++ TombEngine/Math/Objects/Pose.cpp
@@ -1,9 +1,9 @@
 #include "framework.h"
-#include "Math/Containers/Pose.h"
+#include "Math/Objects/Pose.h"
 
-#include "Math/Containers/EulerAngles.h"
-#include "Math/Containers/Vector3i.h"
 #include "Math/Geometry.h"
+#include "Math/Objects/EulerAngles.h"
+#include "Math/Objects/Vector3i.h"
 
 using namespace TEN::Math;
