+++ TombEngine/Math/Objects/Vector3i.cpp
@@ -1,5 +1,5 @@
 #include "framework.h"
-#include "Math/Containers/Vector3i.h"
+#include "Math/Objects/Vector3i.h"
 
 //namespace TEN::Math
 //{
