+++ TombEngine/TombEngine.vcxproj
@@ -165,7 +165,7 @@ CALL gen.bat</Command>
     <ClInclude Include="Game\effects\item_fx.h" />
     <ClInclude Include="Game\effects\lightning.h" />
     <ClInclude Include="Game\GuiObjects.h" />
-    <ClInclude Include="Math\Containers\AxisAngle.h" />
+    <ClInclude Include="Math\Objects\AxisAngle.h" />
     <ClInclude Include="Objects\TR1\Trap\DamoclesSword.h" />
     <ClInclude Include="Scripting\Internal\TEN\Effects\EffectIDs.h" />
     <ClInclude Include="Scripting\Internal\TEN\Misc\ActionIDs.h" />
@@ -186,19 +186,19 @@ CALL gen.bat</Command>
     <ClInclude Include="Game\Lara\lara_jump.h" />
     <ClInclude Include="Game\Lara\lara_overhang.h" />
     <ClInclude Include="Game\Lara\lara_test_structs.h" />
-    <ClInclude Include="Math\Containers\GameBoundingBox.h" />
+    <ClInclude Include="Math\Objects\GameBoundingBox.h" />
     <ClInclude Include="Math\Solvers.h" />
     <ClInclude Include="Objects\game_object_ids.h" />
-    <ClInclude Include="Math\Containers\GameVector.h" />
-    <ClInclude Include="Math\Containers\EulerAngles.h" />
+    <ClInclude Include="Math\Objects\GameVector.h" />
+    <ClInclude Include="Math\Objects\EulerAngles.h" />
     <ClInclude Include="Math\Constants.h" />
     <ClInclude Include="Math\Geometry.h" />
     <ClInclude Include="Math\Legacy.h" />
     <ClInclude Include="Math\Math.h" />
-    <ClInclude Include="Math\Containers\Pose.h" />
+    <ClInclude Include="Math\Objects\Pose.h" />
     <ClInclude Include="Math\Random.h" />
-    <ClInclude Include="Math\Containers\Vector2i.h" />
-    <ClInclude Include="Math\Containers\Vector3i.h" />
+    <ClInclude Include="Math\Objects\Vector2i.h" />
+    <ClInclude Include="Math\Objects\Vector3i.h" />
     <ClInclude Include="Objects\Generic\Object\burning_torch.h" />
     <ClInclude Include="Objects\Generic\Object\polerope.h" />
     <ClInclude Include="Objects\Generic\Object\rope.h" />
@@ -637,7 +637,7 @@ CALL gen.bat</Command>
     <ClCompile Include="Game\collision\floordata.cpp" />
     <ClCompile Include="Game\effects\footprint.cpp" />
     <ClCompile Include="Game\GuiObjects.cpp" />
-    <ClCompile Include="Math\Containers\AxisAngle.cpp" />
+    <ClCompile Include="Math\Objects\AxisAngle.cpp" />
     <ClCompile Include="Objects\TR1\Trap\DamoclesSword.cpp" />
     <ClCompile Include="Scripting\Internal\TEN\Objects\Lara\LaraObject.cpp" />
     <ClCompile Include="Scripting\Internal\TEN\Objects\Room\RoomObject.cpp" />
@@ -667,16 +667,16 @@ CALL gen.bat</Command>
     <ClCompile Include="Game\pickup\pickup_misc_items.cpp" />
     <ClCompile Include="Game\pickup\pickup_weapon.cpp" />
     <ClCompile Include="Game\room.cpp" />
-    <ClCompile Include="Math\Containers\GameBoundingBox.cpp" />
-    <ClCompile Include="Math\Containers\GameVector.cpp" />
-    <ClCompile Include="Math\Containers\EulerAngles.cpp" />
+    <ClCompile Include="Math\Objects\GameBoundingBox.cpp" />
+    <ClCompile Include="Math\Objects\GameVector.cpp" />
+    <ClCompile Include="Math\Objects\EulerAngles.cpp" />
     <ClCompile Include="Math\Geometry.cpp" />
-    <ClCompile Include="Math\Containers\Pose.cpp" />
+    <ClCompile Include="Math\Objects\Pose.cpp" />
     <ClCompile Include="Math\Legacy.cpp" />
     <ClCompile Include="Math\Math.cpp" />
     <ClCompile Include="Math\Random.cpp" />
-    <ClCompile Include="Math\Containers\Vector2i.cpp" />
-    <ClCompile Include="Math\Containers\Vector3i.cpp" />
+    <ClCompile Include="Math\Objects\Vector2i.cpp" />
+    <ClCompile Include="Math\Objects\Vector3i.cpp" />
     <ClCompile Include="Math\Solvers.cpp" />
     <ClCompile Include="Objects\Generic\Object\burning_torch.cpp" />
     <ClCompile Include="Objects\Generic\Object\polerope.cpp" />
